Refactor assembler passes to use utility modules
