+++ Makefile
@@ -1,7 +1,7 @@
 CC = gcc
 CFLAGS = -Wall -Wextra -std=c99 -Isrc
 
-SRCS = main.c parser.c macro.c symbol_table.c symbols.c instructions.c output.c utils.c registers.c src/error.c
+SRCS = main.c parser.c first_pass.c second_pass.c macro.c symbol_table.c symbols.c instructions.c output.c utils.c registers.c src/error.c
 OBJS = $(SRCS:.c=.o)
 
 assembler: $(OBJS)
