+++ main.c
@@ -1,16 +1,16 @@
 // main.c
+#define _POSIX_C_SOURCE 200809L
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 #include "utils.h"
 #include "macro.h"
 #include "parser.h"
 #include "symbol_table.h"
-#include "instructions.h"
 #include "output.h"
 #include "error.h"
+#include "second_pass.h"
 
-static ParsedLine *all_lines = NULL;
-static int         line_count = 0;
 
 /* Read whole file into a lines[] array */
 static bool read_input(const char *fname, char ***out_lines, int *out_n) {
@@ -28,35 +28,6 @@ static bool read_input(const char *fname, char ***out_lines, int *out_n) {
     return true;
 }
 
-/* Second pass: execute each instruction in order (fills cpu.memory) */
-static bool second_pass(CPUState *cpu) {
-    for (int i = 0; i < line_count; i++) {
-        ParsedLine *pl = &all_lines[i];
-        if (pl->type != STMT_INSTRUCTION) continue;
-        /* Dispatch based on opcode string */
-        if      (strcasecmp(pl->opcode,"MOV")==0) exec_mov(pl,cpu);
-        else if (strcasecmp(pl->opcode,"CMP")==0) exec_cmp(pl,cpu);
-        else if (strcasecmp(pl->opcode,"ADD")==0) exec_add(pl,cpu);
-        else if (strcasecmp(pl->opcode,"SUB")==0) exec_sub(pl,cpu);
-        else if (strcasecmp(pl->opcode,"LEA")==0) exec_lea(pl,cpu);
-        else if (strcasecmp(pl->opcode,"CLR")==0) exec_clr(pl,cpu);
-        else if (strcasecmp(pl->opcode,"NOT")==0) exec_not(pl,cpu);
-        else if (strcasecmp(pl->opcode,"INC")==0) exec_inc(pl,cpu);
-        else if (strcasecmp(pl->opcode,"DEC")==0) exec_dec(pl,cpu);
-        else if (strcasecmp(pl->opcode,"JMP")==0) exec_jmp(pl,cpu);
-        else if (strcasecmp(pl->opcode,"BNE")==0) exec_bne(pl,cpu);
-        else if (strcasecmp(pl->opcode,"JSR")==0) exec_jsr(pl,cpu);
-        else if (strcasecmp(pl->opcode,"RED")==0) exec_red(pl,cpu);
-        else if (strcasecmp(pl->opcode,"PRN")==0) exec_prn(pl,cpu);
-        else if (strcasecmp(pl->opcode,"STOP")==0) exec_stop(pl,cpu);
-        else {
-            print_error("Unrecognized opcode in second pass");
-            return false;
-        }
-        cpu->PC++;  /* advance to next memory cell */
-    }
-    return (get_error_count()==0);
-}
 
 int main(int argc, char **argv) {
     if (argc != 2) {
@@ -75,7 +46,6 @@ int main(int argc, char **argv) {
     /* 2. First pass */
     SymbolTable st; init_symbol_table(&st);
     ParsedLine *plarr = malloc(sizeof(*plarr)*flat_n);
-    all_lines = plarr;  line_count = flat_n;
 
     FILE *tmp = tmpfile();
     for(int i=0;i<flat_n;i++) fputs(flat[i],tmp);
@@ -101,7 +71,7 @@ int main(int argc, char **argv) {
     fclose(tmp);
 
     /* 4. Second pass: execute instructions */
-    if (!second_pass(&cpu)) {
+    if (!second_pass(plarr, flat_n, &cpu)) {
         print_error("Second pass failed");
         return 1;
     }
